travis: fix system's ruby 'tlsv1 alert' and brew doctor's HOMEBREW_TRAVIS_CI
1) for the 'tlsv1 alert' on system ruby, it is because old versions of
system ruby give this message. My workaround is to use 'rvm default',
which uses the default version of ruby installed. The only bad thing is
that in this case, because 'brew test-bot' uninstalled every single
installed formulae but rvm default has a ruby version that relies on
brew's openssl@1.1 and such.

The workaround to this workaround is to reinstall openssl@1.1 and such
after every test-bot run.

2) brew doctor has dropped support for Travis CI, meaning that HOMEBREW_TRAVIS_CI
won't trigger the expected behaviour of enabling HOMEBREW_DEVELOPER=1. In order to
reproduce the same behaviour, I have to set

    HOMEBREW_AZURE_PIPELINES=1

See the culprit commit that removed HOMEBREW_TRAVIS_CI:

    https://github.com/Homebrew/brew/commit/61fa59e95bbcfc0585d93788af379ff2eae8594e

and a commit in which I was asking what was the status of Travis CI
support:
    https://github.com/Homebrew/brew/commit/59c858a8e8362e7cdc063e107d0360be9dc85d0e